set lazy load to eager (temp solution)
