+++ components/ChakraNextImage.js
@@ -20,6 +20,7 @@ function Img(props) {
       src={src}
       alt={alt}
       srl_gallery_image="true"
+      loading="eager"
       placeholder={!regex.test(src) && "blur"}
     />
   );
