+++ components/GalleryWrapper.js
@@ -1,0 +1,26 @@
+import SimpleReactLightbox, { SRLWrapper } from "simple-react-lightbox-pro";
+
+export default function GalleryWrapper(props) {
+  const options = {
+    settings: {
+      overlayColor: "rgba(30, 30, 30, 1)",
+    },
+    buttons: {
+      showDownloadButton: false,
+      showAutoplayButton: false,
+      showThumbnailsButton: false,
+    },
+
+    thumbnails: {
+      showThumbnails: false,
+    },
+    caption: {
+      showCaption: false,
+    },
+  };
+  return (
+    <SimpleReactLightbox>
+      <SRLWrapper options={options}>{props.children}</SRLWrapper>
+    </SimpleReactLightbox>
+  );
+}
